worker: move the userns holder into its own module
The Linux build agent creates its user namespace with the same
unshare-and-pause holder child the sandboxd client uses. Lift
UsernsHolder and the pidfd/ns-path helpers out of sandboxd.rs into
worker/userns.rs so both can share them.
